feat: components refactoring, reconfigurable banners and buttons
