--- blockwarriors-next/src/app/components/DescriptionText.jsx
@@ -1,16 +1,0 @@
-import React from "react";
-
-export default function DescriptionText(props) {
-  return (
-    <div>
-      <h2 className="text-center text-4xl">{props.title}</h2>
-      <p className="text-center text-xl">
-        Nisi veniam dolor ea sint do veniam sit. Cupidatat laborum exercitation
-        adipisicing ullamco exercitation occaecat occaecat sint. Dolore minim
-        irure ullamco esse incididunt proident do sint id duis excepteur do
-        nisi. Esse ex consequat in Lorem non. Cillum ex labore nisi deserunt
-        ipsum proident eiusmod et est commodo ullamco.
-      </p>
-    </div>
-  );
-}
