+++ blockwarriors-next/src/app/dashboard/(components)/TournamentProgress.tsx
