+++ blockwarriors-next/src/app/login/(components)/LoginForm.tsx
@@ -1,0 +1,113 @@
+'use client';
+
+import { useState } from 'react';
+import { Loader2 } from 'lucide-react';
+import Link from 'next/link';
+
+export default function LoginForm() {
+  const [isLoading, setIsLoading] = useState(false);
+  const [formData, setFormData] = useState({
+    username: '',
+    password: '',
+  });
+
+  const handleSubmit = async (e: React.FormEvent) => {
+    e.preventDefault();
+    setIsLoading(true);
+    // Authentication logic will be implemented later
+    setIsLoading(false);
+  };
+
+  return (
+    <form onSubmit={handleSubmit} className="space-y-3">
+      <div>
+        <label
+          htmlFor="username"
+          className="block text-sm font-medium text-gray-700 dark:text-gray-300"
+        >
+          Username
+        </label>
+        <input
+          type="text"
+          id="username"
+          name="username"
+          disabled
+          value={formData.username}
+          onChange={(e) =>
+            setFormData({ ...formData, username: e.target.value })
+          }
+          className="mt-1 block w-full px-3 py-1.5 bg-gray-100 dark:bg-gray-700 border border-gray-300 dark:border-gray-600 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 disabled:opacity-75 disabled:cursor-not-allowed"
+          required
+        />
+      </div>
+
+      <div>
+        <label
+          htmlFor="password"
+          className="block text-sm font-medium text-gray-700 dark:text-gray-300"
+        >
+          Password
+        </label>
+        <input
+          type="password"
+          id="password"
+          name="password"
+          disabled
+          value={formData.password}
+          onChange={(e) =>
+            setFormData({ ...formData, password: e.target.value })
+          }
+          className="mt-1 block w-full px-3 py-1.5 bg-gray-100 dark:bg-gray-700 border border-gray-300 dark:border-gray-600 rounded-md shadow-sm focus:outline-none focus:ring-indigo-500 focus:border-indigo-500 disabled:opacity-75 disabled:cursor-not-allowed"
+          required
+        />
+      </div>
+
+      <div className="flex items-center justify-between text-sm">
+        <div className="flex items-center">
+          <input
+            id="remember-me"
+            name="remember-me"
+            type="checkbox"
+            disabled
+            className="h-4 w-4 text-indigo-600 focus:ring-indigo-500 border-gray-300 rounded disabled:opacity-75 disabled:cursor-not-allowed"
+          />
+          <label
+            htmlFor="remember-me"
+            className="ml-2 text-gray-700 dark:text-gray-300"
+          >
+            Remember me
+          </label>
+        </div>
+
+        <Link
+          href="/forgot-password"
+          className="font-medium text-indigo-600 hover:text-indigo-500 opacity-50 pointer-events-none"
+        >
+          Forgot password?
+        </Link>
+      </div>
+
+      <button
+        type="submit"
+        disabled
+        className="w-full flex justify-center py-1.5 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white bg-indigo-600 hover:bg-indigo-700 focus:outline-none focus:ring-2 focus:ring-offset-2 focus:ring-indigo-500 disabled:opacity-50 disabled:cursor-not-allowed"
+      >
+        {isLoading ? (
+          <Loader2 className="w-4 h-4 animate-spin" />
+        ) : (
+          'Sign in'
+        )}
+      </button>
+
+      <p className="text-center text-sm text-gray-600 dark:text-gray-400">
+        Don&apos;t have an account?{' '}
+        <Link
+          href="/register"
+          className="font-medium text-indigo-600 hover:text-indigo-500 opacity-50 pointer-events-none"
+        >
+          Sign up
+        </Link>
+      </p>
+    </form>
+  );
+}
